Implementation/cookbook/cmdpm/cmdpm/utils/config.py
import yaml
import os

def load_config(config_path):
    """
    Load configuration from a YAML file.
    
    Args:
        config_path (str): Path to the YAML configuration file.
    
    Returns:
        dict: Loaded configuration.
    """
    with open(config_path, "r") as f:
        config = yaml.safe_load(f)
    return config

def save_config(config, config_path):
    """
    Save configuration to a YAML file.
    
    Args:
        config (dict): Configuration to save.
        config_path (str): Path to save the YAML file.
    """
    os.makedirs(os.path.dirname(config_path), exist_ok=True)
    with open(config_path, "w") as f:
        yaml.dump(config, f)